pyproject.toml
[tool.poetry]
name = "ATTACK-RISK"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
pgmpy = "^0.1.25"
polars = "^0.20.31"
mitreattack-python = "^3.0.6"
stix = "^1.2.0.11"
matplotlib = "^3.9.1"
setuptools = "^70.3.0"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.4"

[tool.mypy]
warn_return_any = true
warn_unused_configs = true
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
